projects/altclaw/analytics/requirements.txt
google-analytics-data==0.18.12
google-auth==2.37.0
google-auth-oauthlib==1.2.1
